Start a new Element module for pure widgets.
