+++ test/Main.hs
@@ -2,7 +2,7 @@ module Main (main) where
 
 import Test.Tasty
 
-import CoreSpec qualified
+import ElementsSpec qualified
 import DecoderSpec qualified
 import PlaylistSpec qualified
 import StyleSpec qualified
@@ -10,7 +10,7 @@ import TextSpec qualified
 
 main :: IO ()
 main = defaultMain $ testGroup "hmp3-ng"
-    [ CoreSpec.tests
+    [ ElementsSpec.tests
     , DecoderSpec.tests
     , StyleSpec.tests
     , PlaylistSpec.tests
